{{ cookiecutter.project_name }}/backend/requirements.txt
pypandoc
inotify
gunicorn
Click
Flask
Flask-Cors
Flask-Login
Flask-Bcrypt
flask_wtf
itsdangerous
Jinja2
MarkupSafe
six
Werkzeug
pymongo
pyyaml
requests
pandas
xlrd
openpyxl
python-dotenv
pandas
numpy
attrs
mypy
cattrs

